add examplemodal
